Tie exercise Jwq more directly to the physical context
I'm proposing a couple of changes to the wording of the exercise with permid Jwq (it's exercise 5 in the current html build). The context of this exercise is that a moving particle has its velocity given by a particular graph with some areas indicated. I embroidered a couple of the questions with some more physical language (e.g., when is the particle moving forward / backward) to give students a stronger nudge to make those connections.